+++ impl/src/main/java/uk/ac/manchester/cs/AcceptHeaderBuilder.java
@@ -30,12 +30,13 @@ public class AcceptHeaderBuilder {
         // The map will contain all mime types with their position in all lists mentioning them; the
         // smallest position first
         for (int i = 0; i < mimes.size(); i++) {
-            map.computeIfAbsent(mimes.get(i), k -> new TreeSet<>()).add(i + 1);
+            map.computeIfAbsent(mimes.get(i), k -> new TreeSet<>()).add(Integer.valueOf(i + 1));
         }
     }
 
     private static String tostring(Map.Entry<String, TreeSet<Integer>> e) {
-        return String.format("%s; q=%.1f", e.getKey(), 1D / e.getValue().first());
+        return String.format("%s; q=%.1f", e.getKey(),
+            Double.valueOf(1D / e.getValue().first().intValue()));
     }
 
     private static int compare(Map.Entry<String, TreeSet<Integer>> a,
